Renamed Checkbox to Task, fixed issue with unique identifier in data.js
