+++ src/layouts/Main.vue
@@ -7,13 +7,14 @@
 		<div v-else>
 			<span>My tasks :</span>
 
-			<ul v-for="task in sortTasks()">
-				<Checkbox
-					:task="task"
-					:key="task.date"
-					@check="updateTask(task, true)"
-					@uncheck="updateTask(task, false)"
-				/>
+			<ul v-for="task in sortedTasks">
+				<li :key="task.date">
+					<Task
+						:task="task"
+						@check="checkTask"
+						@uncheck="uncheckTask"
+					/>
+				</li>
 			</ul>
 
 			<label id="hide-completed">
@@ -25,26 +26,25 @@
 </template>
 
 <script setup>
-import { ref } from 'vue';
+import { ref, computed } from 'vue';
 import tasks from '../data';
-import Checkbox from '../components/Checkbox.vue';
+import Task from '../components/Task.vue';
 
 const tasksRef = ref(tasks.value);
 const hideCompleted = ref(false);
 
-const updateTask = (task, completed) => {
-	const index = tasksRef.value.findIndex((t) => t.date === task.date);
-
-	if (index !== -1) {
-		tasksRef.value[index].completed = completed;
-		tasks.value = tasksRef.value;
-	}
-
-	console.log(tasks.value[0]);
+const checkTask = (task) => {
+	task.completed = true;
+	tasksRef.value = [...tasksRef.value];
 };
 
-const sortTasks = () => {
-	const sortedTasks = tasks.value.toSorted((a, b) => {
+const uncheckTask = (task) => {
+	task.completed = false;
+	tasksRef.value = [...tasksRef.value];
+};
+
+const sortedTasks = computed(() => {
+	const sortedTasks = tasksRef.value.slice().sort((a, b) => {
 		if (a.completed === b.completed) {
 			return 0;
 		} else if (a.completed) {
@@ -54,12 +54,12 @@ const sortTasks = () => {
 		}
 	});
 
-	if (hideCompleted.value === true) {
+	if (hideCompleted.value) {
 		return sortedTasks.filter((t) => !t.completed);
 	}
 
 	return sortedTasks;
-};
+});
 
 const hideCompletedTasks = () => {
 	return (tasks.value = tasks.value.filter((task) => !task.completed));
@@ -82,6 +82,7 @@ p {
 
 ul {
 	padding-left: 0.5rem;
+	list-style: none;
 }
 
 label#hide-completed {
